Rename place picker class 3
