+++ app/src/main/AndroidManifest.xml
@@ -39,7 +39,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".favourites.AddFavouriteActivity"
+            android:name=".favourite.AddFavouriteActivity"
             android:label="@string/title_add_favourite"/>
     </application>
