+++ app/src/main/res/layout/activity_add_favourite.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".PlacePickerActivity">
+    tools:context=".favourite.AddFavouriteActivity">
 
     <org.osmdroid.views.MapView
         android:id="@+id/map"
